Move server-related files
